tast-tests: arc: Add arc.Reboot and arc.Reboot.vm tests
This is a port of cheets_ContainerReboot autotest. It checks whether
Android can be repeatedly rebooted. Reboot command is executed via ADB.
The success of reboot is checked by seeing whether ARC init PID is
changed after reboot.
When ARCVM is enabled, the test also checks whether vsock CID of ARCVM
is changed after reboot. We export a part of the function arc.New() as
arc.WaitAndroidBoot() and use it to wait for Android boot and
re-establishment of ADB connection.

BUG=b:142574527
TEST=~/trunk/src/platform/tast/fast_build.sh -T
TEST=tast run <DUT-IP-eve> arc.Reboot/arc.Reboot.vm

Change-Id: If5a9786a5d7ff1eb2b929898fc7ced9e64dcb06a