tast-tests: Add arc.Reboot/arc.Reboot.vim test
This is a port of cheets_ContainerReboot autotest. It checks whether
Android can be repeatedly rebooted. Reboot command is executed via ADB.
Whether reboot succeeds is checked by seeing the change in PID.
In order to re-establish the connection to ADB, we have exported the
function connectADB() in src/chromiumos/tast/local/arc/adb.go.

BUG=b:142574527
TEST=tast run <DUT-eve> arc.Reboot/arc.Reboot.vm

Change-Id: If5a9786a5d7ff1eb2b929898fc7ced9e64dcb06a