Get metrics consent via metrics_client
Remove flakiness around setting metrics consent.

When we set metrics consent in Chrome, it does (at least) two things --
it touches the legacy consent file (/home/chronos/Consent To Send Stats)
and it updates the policy file. The first is done instantly; the second
is done asynchronously by queuing a task on a separate thread which then
makes a dbus call to session manager.

Before this change, the tast tests only checked that the legacy consent
file had been created before proceeding with the test. But
crash_reporter checks the policy files if they exist in preference to
checking for the legacy consent files. This created a race condition.

To avoid that, have the tests rely on the same consent checking code
that crash_reporter relies on, instead of our own checks. "metrics_client
-c" already runs the check we want, so just invoke that.

BUG=chromium:1041062
TEST=tast tested ui.ChromeCrash* platform.*Crash* and crash.*.
Everything passed except some unrelated issues with
ui.ChromeCrashLoop.breakpad and platform.UserCrash

Change-Id: I103d637230123bea644d1f57f0cbf9fe7e2ab291