tast/video: Add support for AMD debug graphics object count
This CL adds the graphics.Backend implementation for AMD Adreno
and uses it for the video.MemCheck test cases. It also removes
the Supports...() in favour of a getBlaBackend() and adds a
mini factory function so that mem_check.go() can avoid knowing
which backend is in use.

Potential systems with both amd and intel backends are not
supported and make the test fail.

BUG=chromium:1047514
TEST=tast run video.MemCheck.* on grunt and atlas

Change-Id: I74b2652d0219a67ae139d0fcf7a5e36816f4c0cb