Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I see. I'll leave the sheriff documentation with the 'better' method then, although the length of time involved is unfortunate - this CL took many hours to run the CQ and ultimately failed on two platforms, by which time I was no longer build sheriff!

I agree. Especially for folks coming from the Chrome browser side, the Chrome OS CQ is more heavyweight by nature given the fact that we're building an entire OS image for tens of boards in every CQ run, and then running on every one of these boards.

The right way to do this would have been to just bypass the CQ in this case.

Patch-set: 3