cryptohome: Teach cryptohome package about namespaces.
Soon we'll enable mounting Guest sessions in non-root mount namespaces
as per
https://docs.google.com/document/d/1tvloiVoGmrkD8QYV4SutFuWV8GGb1QXDB3kvAzKfMwc/edit.

Update the Tast cryptohome module so that it can find mounts specific
to a process. This allows finding mounts in any namespace by listing
the mounts of a process in that namespace. Currently, the
disk.Partitions() code only searches for mounts in the root mount
namespace.

(Also, "Partitions" is a bad name for what the function returns.)

By allowing the cryptohome module to retrieve mounts as seen from any
process we let the test pass before and after we enable non-root mount
namespaces, because the test will always check in the right namespace
(by looking at the namespace for the 'cryptohome-namespace-mounter'
process.)

BUG=chromium:985492
TEST=tast run IP cryptohome.GuestSession,
TEST=tast run IP hwsec.LoginGuest
TEST=Pass with and without https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2063039.

Change-Id: Iaec3e6d259ff2d031c81910fffc7c7f0d74757e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2064729
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>