GuestSession: Remove superfluous checks.
The MountGuest function will actually check that the Guest vault was
successfully mounted, so the checks here are 1. unnecessary and 2.
break encapsulation. The test shouldn't know what the ephemeral mount
path is, that knowledge should be encapsulated in the cryptohome
module.

Even without the extra checks this is a very useful test to ensure
Guest sessions don't break, especially since we're making changes in
this area.

BUG=chromium:985492
TEST=Still passes.

Change-Id: I20fc2cf2dad1c6159141ff479307bfd820e568e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2064731
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>