tast/video: Promisify data/video.html and play/play.go
This CL converts the data/video.html and play/play.go appropriate
functions to Promises. This makes the code significantly smaller
and easier to modify, which is needed in the next CL to play the
videos for a while while counting the memory consumption.

Incidentally the new code has better error reporting, because we
allow the play() error to bubble up to Tast (tried it e.g. by
forcing a faulty |video_source_path|).

No functional changes intended, only refactoring.

BUG=chromium:1047514
TEST=tast run video.Play.* video.Seek.* on atlas

Change-Id: Idc6801fa9fe670ab3a60920ba11d0a663238233c