tast/video: Promisify data/video.html and play/play.go
This CL converts the data/video.html and play/play.go appropriate
cases to Promises. This makes the code significantly smaller and
easy to modify, which is needed in the next CL to play the videos
for a while while counting the memory consumption.  Incidentally
the new code has better error reporting, because of the reject()
syntax.

No functional changes intended, only refactoring.

BUG=chromium:1047514
TEST=tast run video.Play.* video.Seek.* on atlas

Change-Id: Idc6801fa9fe670ab3a60920ba11d0a663238233c