Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> arc1 is the equivalent of arc0 for arcvm - maybe this test needs to be disabled for arcvm instead?

Yes, this test is disabled for ARCVM.
But the way the test work is it finds the equivalent interface for guest and host (e.g. "eth0"). This breaks the test as arc1 is not available on the host.

I'll update the comment.

Patch-set: 1
Reviewer: Gerrit User 1346383 <1346383@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>