Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > arc1 is the equivalent of arc0 for arcvm - maybe this test needs to be disabled for arcvm instead?
> 
> Yes, this test is disabled for ARCVM.
> But the way the test work is it finds the equivalent interface for guest and host (e.g. "eth0"). This breaks the test as arc1 is not available on the host.
> 
> I'll update the comment.

Do you mean because if ARCVM is run first then it leaves arc_br1 and arc1 lying around? I don't see how arc1 could be not available on the host but available on the guest if the guest in this case is arc++. What am I missing?

Patch-set: 1