Update patch set 7
Patch Set 7: Code-Review-1

(2 comments)

I think this is a bad idea. If you really want the device list, just ask Shill for it.

BTW, I'm not really sure what the problem with the original message ("unable to find WiFi interface") is -- it's telling you it couldn't find the WiFi interface. And unless the test code is buggy (I don't think it is?) the problem is that...the WiFi interface is missing.

Patch-set: 7
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1