Update patch set 10
Patch Set 10:

(2 comments)

> + Matthew, as he's dealt with some of this in Autotest (esp. bridging aspects, which we don't handle yet, but will need to eventually)
> 
> I realize some of my comments below are making armchair suggestions on significant refactoring. I partly make these suggestions because
> (a) Autotest has a garbage-pile of intertwined stuff in LinuxRouter, with no logical separations
> (b) you've started to unwind some of this in the port, and I think it's a great start and
> (c) I think we could do a little more
> 
> Let me know if I'm off-base, or if you want to have a quicker (not on Gerrit) discussion about these things. I really want to not slow this down too much more.

Refactored a little bit. Moving interface related logic to HostAPHandle looks nice for future extension.

After a few looks on the bridging, I think the difference on DHCP part is that the server runs on different interface (and corresponding interface configuration, of course):
  1. start_dhcp_server runs dhcp server on bridge interface instead if available.
  2. setup_bridge_mode_dhcp_server runs dhcp server on veth pair and add the other side to bridge.
For (1), I guess it can be done by extending the start/stop logic in HostAPHandle.
For (2), I guess we can further extract dhcp setup from HostAPHandle so that it can share. (but setup_bridge_mode_dhcp_server is only used by apmanager_SimpleConnect and I didn't see any result on stainless, so I'm not sure if it's important.)

Patch-set: 10