Update patch set 2
Patch Set 2:

> - Move OWNERS file one dir deeper, i.e. tast-tests/android/ArcGamepadTest/OWNERS, because it is expected that more .apks come here from various teams.

The intention here is that we want to be the owners of all Arc*Test that will be added in the future, so android/OWNERS is the right location.

* More apks are expected to be moved from vendor/google_arc of the pi-arc branch. These files had to go through reviews by the people in the OWNERS file when they were under vendor/google_arc, and we already have some familiarity with them.
* These files are Android projects which are quite different from other parts of Tast which are golang based.

Based on that, I believe it is reasonable to have different set of reviewers for this directory. For the same reason, I thought set noparent might be suitable here.

Given that, I would like to hear opinions from hidehiko@ and nya@ again. I understand there are disadvantages as well e.g. reviewer fragmentation, so it is also fine if you think we should avoid this.

Patch-set: 2