Update patch set 2
Patch Set 2:

CQ for this CL appears to be stuck on unscheduled (pending) tasks.  I haven't checked (and won't at this hour) if that's related to the breakage this revert is intended to fix, but I assert that either way force-submitting this is correct.

Patch-set: 2
Reviewer: Gerrit User 1291807 <1291807@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>