tast/graphics: fix: Adreno backend is actually a generic
crrev.com/c/2063028 landed the necessary dri.go Backend for Adreno,
and left a provision that both i915 and Adreno backends should not
be present at once. This is my bad: the Adreno backend is actually
the generic Linux one, so modern kernels also have that present on
Intel systems (⌒_⌒;) -- making the test fail.

This CL then renames s/Adreno/Generic/ and avoids falling when
both are present, simply choosing the "generic" one.

BUG=chromium:1047514
TEST=tast run video.MemCheck.*  on atlas and grunt (kasumi)

Change-Id: If3a07ba4ad419834dae749d2c15775ee51d4d298