shill: use IsDBusError to skip error of removed device
Only skip the dbus error DBusErrorUnknownObject in DeviceByName. Also,
as we don't have an implicit GetProperty call in NewDevice now, it will
not fail on removed object path (in dbusutil.Connect, it only checks
availability of the service). We can just return the error if we fail
on NewDevice.

Also add some more documentation for the related behavior.

BUG=chromium:1047568
TEST=fast_build.sh -T
TEST=create a local Tast test with test body:
```
func POCLocal(ctx context.Context, s *testing.State) {
        dev, err := shill.NewDevice(ctx, "/invalid/path")
        if err != nil {
                s.Fatal("NewDevice error:", err)
        }
        _, err = dev.GetProperties(ctx)
        if err != nil {
                if !dbusutil.IsDBusError(err, dbusutil.DBusErrorUnknownObject) {
                        s.Fatal("GetProperties error:", err)
                }
        } else {
                s.Fatal("expect error")
        }
}
```
Run this test to make sure we're skipping the error correctly.

Change-Id: I47bf2e80877eb4f1935982eb1efc3bb74022e807