Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > A mild concern that I have is that 720p is closer to user experience than 240p. But if you are mostly concerned about being lightweight then sure.
> 
> A few questions:
> 1) Is it likely that we'd have a leak with 240p but not with 720p?

Hard to tell. But the probabilities should be conditional. Assume users found a leak, why did the test not tell us? Well, the answer is we tested something different than the users. How different?

> 2) Is it easier to detect leaks with 240p or 720p?

For the test or for a user? A user may experience leaks earlier with 720p. Miguel's code fortunately should not care, as it tries to account for the right sizes.

> 3) When we say "lightweight", what is the bottleneck in testing that distinguishes between 240p and 720p? (I would guess memory is the main distinction, not length, but would not expect memory is a bottleneck)

I would think speed/races in the code are the main difference.

But I thought this change is not mainly about the memory test?

Any case, I think it is fine to use 240p here. We don't capture the video to compare (yet). But I keep thinking there is benefit in exercising all sizes.

Patch-set: 1