Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > (1 comment)
> > > 
> > > A mild concern that I have is that 720p is closer to user experience than 240p. But if you are mostly concerned about being lightweight then sure.
> > 
> > A few questions:
> > 1) Is it likely that we'd have a leak with 240p but not with 720p?
> 
> Hard to tell. But the probabilities should be conditional. Assume users found a leak, why did the test not tell us? Well, the answer is we tested something different than the users. How different?
> 

It's impossible to tell the nature of an unknown leak :-) but those we know 
could all be traced to amount of resources, not their sizes.

The way we count those resources in the current tests is via framebuffer-bound 
DmaBufs, which is very targeted, versus amount-of-allocated graphics "stuff" 
which is not: we found a leak this way but it's hard to make a robust test out
of that (there are other teams fortunately monitoring OOMs etc).

> > 2) Is it easier to detect leaks with 240p or 720p?
> 
> For the test or for a user? A user may experience leaks earlier with 720p. Miguel's code fortunately should not care, as it tries to account for the right sizes.
> 
> > 3) When we say "lightweight", what is the bottleneck in testing that distinguishes between 240p and 720p? (I would guess memory is the main distinction, not length, but would not expect memory is a bottleneck)
> 
> I would think speed/races in the code are the main difference.
> 
> But I thought this change is not mainly about the memory test?

Correct, this test fixes the fact that MSE and non-MSE ("normal") video playback
have fundamentally different mechanisms: the former runs until onended, which
works fine because the bear videos are 2s long, while the latter runs until
1.5s have passed, via monitoring the onupdate video event (which is inaccurate
by design BTW). I don't expect improvement because the bear video is of a smaller
resolution.

This change makes it possible to play a video for its entire length
while monitoring the amount of allocated framebuffers, which I'd like to
land as well.

> 
> Any case, I think it is fine to use 240p here. We don't capture the video to compare (yet). But I keep thinking there is benefit in exercising all sizes.

Probably - but the video.Play.* were supposed to be the easiest verification
of the playback functionality. We could always add video.Resolutions or
some such.

Patch-set: 1
Reviewer: Gerrit User 1113991 <1113991@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>