tast-tests: Add check for ARCVM status in preconditions
Whereas each precondition is either ARC Container specific or ARCVM
specific, currently we do not have a check for whether ARCVM is actually
enabled/disabled in the preconditions.
This CL adds such a check.
The check is done using the function VMEnabled(), and an error is thrown
if the result is inconsistent with the precondition name. We assume the
preconditions specifying ARC Container have names starting from "arc_",
whereas those specifying ARCVM have names starting from "arcvm_".

BUG=b:150428197
TEST=~/trunk/src/platform/tast/fast_build.sh -T
TEST=Confirm all the existing preconditions pass the check
TEST=Force an error and see whether error messages look pretty

Change-Id: If5217e93c0e57a94e095837da139c9b09a4943e1