tast-tests: Add check for ARCVM status in preconditions
Whereas each precondition is either ARC Container specific or ARCVM
specific, currently we do not have a check for whether ARCVM is actually
enabled/disabled in the preconditions.
This CL adds such a check.
The check is done using a new flag useARCVM added to the preImpl struct.
If the return value of VMEnabled() called after Chrome is started is
inconsistent with the flag, the precondition throws an error.
Also, the flag is used to determine whether "--enable-arcvm" is added to
extraArgs for Chrome.
Note that there are some tests that are specific to either ARC Container
or ARCVM, but at the same time do not use any preconditions, e.g.,
arc.Boot/arc.Boot.vm. This CL does not provide a way for these tests to
check ARCVM status.

BUG=b:150428197
TEST=~/trunk/src/platform/tast/fast_build.sh -T
TEST=Confirm all the existing preconditions pass the check
TEST=Force an error and see whether error messages look pretty

Change-Id: If5217e93c0e57a94e095837da139c9b09a4943e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2078637
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Auto-Submit: Youkichi Hosoi <youkichihosoi@chromium.org>