tast/video: use new video_overlays sw dependency for MemCheck
video.MemCheck.* tests depend on the video decoder graphics buffers
being marked as promotable to overlay, which only happens on boards
where the DRM subsystem supports NV12 as input format for the planes
(and that format is used). crrev.com/c/2080598 introduces the
video_overlays software dependency, and this CL uses it.

BUG=chromium:1047514
TEST=tast run ${IP} video.MemCheck.* on samus and atlas
samus skips and atlas runs correctly and as expected

Change-Id: I84d2f094659d2e2a0cbe0b059bc01a6c93c1e0e2