consolidate Size and TouchCoord conversion
In the previous CL, I created the definition of Size struct but
forgotten to make the transition on existing types. This CL makes
the consolidation of arc.Size and crostini.Size to coords.Size.

Through this task, I've realized that sometimes size is used for
converting with input.TouchCoord. Thus this CL also moves
ash.TouchCoordsConverter to coords package and use it everywhere
else.

BUG=chromium:1055030
TEST=fast_build.sh -T, run affected tests

Change-Id: Ie30695d7974be78bf9764f4bf24a50f087e59e2a