tast-test: Skip elm/hana from running network.UdevRename.
Because a known issue in elm/hana, rebinding the WiFi kernel module
could cause kernel panic/reboot. If the device is not rebooted, it has a
broken WiFi and affects the upcoming WiFi related tests. Tast now
provides HardwareDeps to skip certain model for a test, we can use it to
skip elm/hana from running UdevRename. elm/hana can be added back once
the kernel is upreved.

BUG=chromium:1025091
TEST=tast run $dut network.UdevRename on hana/elm and expect the test skips.

Change-Id: I40d9e68a881d887c14c4ef7bf037d7f774768da7