tast-test: Skip elm/hana from running network.UdevRename.
network.UdevRename constantly fails on elm/hana because of a known issue
that the wifi driver unbind is racing with firmware dump that the dumper
is accessing memory that was already freed by the unbind operation. See
https://crbug.com/1025091#c6.

Leave the test running on elm/hana may cause other tests flaky. It is
because the issue mentioned above does not always trigger a reboot. When
it is not rebooted, the device does not have a WiFi interface hence WiFi
dependent tests are failed.

Tast now provides HardwareDeps to skip certain model for a test, we can
use it to skip elm/hana from running UdevRename. elm/hana can be added
back once the kernel on elm/hana is upreved.

BUG=chromium:1025091
TEST=tast run $dut network.UdevRename on hana/elm and expect the test skips.
Test skip log:
network.UdevRename  [ SKIP ] PlatformId matched with skip-on list

Verified on
elm-release/R82-12940.0.0
elm-release/R81-12871.30.0
elm-release/R80-12739.91.0

Change-Id: I40d9e68a881d887c14c4ef7bf037d7f774768da7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2081555
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>