accessibility: avoid chrome.Conn/chrome.TestConn crossover
The chromeVoxConn connects to a different extension than the
TestConn does and so it uses a regular Conn, but it uses
ui.NewNode, which is defined to take a TestConn. Thread the
TestConn through so it can be properly passed to ui.NewNode.

BUG=chromium:1043590
TEST=tast run arc.AccessibilityEvent arc.AccessibilityTree \
  arc.AccessibilitySpeech

Change-Id: I9fdd10d6b280234c2f3dac7fd0f2182800343490
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2083848
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>