Update patch set 1
Patch Set 1:

Err, it seems like we have a separate issue -- the TestConn type seems to be correct because that object gets passed to e.g. ash.MouseClick, which is communicating with the autotestPrivate API. So I'm not entirely sure where to go from here, in that case.

Patch-set: 1
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>