Update patch set 7
Patch Set 7:

Sorry about the changes, they were not from changes to lower commits. Cleaning up the DBus connection properly is exactly the sort of thing Setup is made for, so I put the cleanup into its own setup action. The API has been stable for a few weeks now, I just changed the commit to use it.

I was just careless with the cleanup in the initial implementation.

I'm new to Google and Go, so I was not expecting code review to be so involved.

Patch-set: 7