Add FAFT configs to Tast
This CL copies the firmware-testing configs from Autotest into Tast,
a necessary step toward implementing SwitchBootMode.

It is not ideal to have duplicated code. However, the long-term solution
for storing firmware-testing configs is not ready for broad use. This is
a temporary solution so that we can continue feature development.

For more information, see go/cros-fw-testing-configs.

BUG=chromium:993062
TEST=None

Change-Id: I95e579296ce713461c6fee8b8d11a4f5571503c7