tast-tests: Add test to verify Drallion power key behavior
On Drallion360, the power button is on the keyboard rather than the side
of the device.  This makes inadvertent keypresses while in tablet mode
more common. To account for this, the power button requires a long press
(>1.5s) beore bringing up the power-down menus. This is the behavior
that we must verify with this test.

BUG=b:149035007
TEST=Run test on lab DUT, verify expected behavior:
tast -verbose run -var servo={servo_addr}.cros:{servo_port} {dut_name}.cros power.DrallionTabletPower

Cq-Depend: chrome-internal:2761210
Change-Id: I84df7f6792d16e8bbb81701f182c3e3f1d495eb4