Update patch set 4
Patch Set 4:

Thanks for this change! Seems very helpful.

It does add a good amount of boilerplate, though. Could you instead add a method to fixture.go that tears down or causes the test to error, and doesn't return anything? That way we avoid code duplication. (I'd also be fine with changing TearDown's behavior to do this; either way.)

Patch-set: 4