tast/video: avoid Intel platforms without hardware overlays support
Some Intel platforms have a (relatively) new kernel, and one that
supports DRM Atomic (so the sw dep "video_overlays") is satisfied;
however we don't bind video to framebuffers, and the MemCheck test
fails, because it counts 0 such framebuffers.

crbug.com/1057870 has a permanent solution to it, via surfacing the
chipset and using it for sw dependency definition; in the meantime
and as stopgap solution, this CL adds a SupportsHardwareOverlays()
function that returns true if we're on AMD or NV12 is supported.

BUG=chromium:1047514,b:149958753
TEST=tast run $IP video.MemCheck.* on grunt, coral(babymega), kohaku.

Change-Id: Idd6f4cf6c6063b74ea40fcd0ce995a59325471a0