Fix LauncherPageSwitchPerf failure on krane
Looks like the first page of the apps-grid might have huge blank
at the bottom since the default apps are arranged there, and
mouse events in such blank area don't cause the page-scrolling.

This CL simply switches to the second page before conducting the
drag operation to prevent this type of error to happen. Because
it is already in the second page, it doesn't have to conduct
the drag-up operation twice.

BUG=chromium:1058591
TEST=tast run ui.LauncherPageSwitchPerf.clmashell_mode on krane

Change-Id: Id64727c9ae8adbb7983ae181949d7f9abd90a9ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2088621
Commit-Queue: Jun Mukai <mukai@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>