arc: Explicit expectation for ARC-P boards in BuildProperties.
FirstApiLevel is expected to be equal to the current api level,
unless explicitly overriden in the test code. Until now, the
implicit expectation was okay for the boards launched with P
(i.e., those expected to have firstApiLevel=P), but for the
future development for switching them to newer desserts, we
need to declare we expect P first api level for them.

BUG=chromium:1058793
TEST=arc.BuildProperties on Nocturne.

Change-Id: I461f1f583e76201dcee98b653808a4319c3e6e5b