Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > hidehiko@ PTAL -- alternatively we could hold on if crbug.com/950346 is landing 
> >                   support for SOCs soon.
> 
> Skip the test on runtime is what we ask test authors to avoid.
Correct and agree, this was never meant to be a full solution.

> crbug.com/950346: the framework of hwdep is already available. Remaining for this is crbug.com/1057870 (ie. fill SOC field). Would you mind to use that approach?

If landing support for hwdeps is already in the radar then it makes
sense to wait for it. Will abandon this CL, let's continue in
crbug.com/950346 and crbug.com/1057870. 

Thanks reviewers!

Patch-set: 1
Reviewer: Gerrit User 1113991 <1113991@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0