Fix the failures of ui.LauncherDragPerf on some devices
As I noted in the bug, the problem happens because I was assuming
there is only one app icon in the shelf for this test. This is
not the case when it runs with ui.HotseatScrollPerf, which adds
many apps. To conduct launcher-dragging properly, the starting
point at the bottom should be somewhere in the gap among app icons.

This CL searches the app-icons, and picks up the point which should
cause the launcher to appear:
* if there are many app icons, simply pick up the gap between the
  first icon and the second one.
* if there's only one icon (chrome icon), pick up the right of
  that icon.
* there shouldn't be no-icons scenario.
* X position at the top should be 1/4 width of the screen; it might
  coincide with the indicator icon of the shelf and might cause
  failures of drag-to-close.

BUG=chromium:1058216
TEST=tast run <IP> ui.HotseatScrollPerf ui.LauncherDragPerf

Change-Id: I21f5c292247b717b05f38ccd5814103276322afc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2092694
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>