Update patch set 8
Patch Set 8: Code-Review+1

Thanks for the clarifications, Victor and Shuhei.

LGTM, given that we're assuming that no test leaves the device in an enrolled state, so that this CL fixes the remaining bad case when a test does unenrollment but forgets to remove the device policy file (BTW, it'll be good to file a bug to fix this in autotest tests as well).

Patch-set: 8
Reviewer: Gerrit User 1126309 <1126309@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1