Wait for owner.key file to become present.
If we don't wait for the device policy file to be present after setting
consent, we'll have a brief period after SetConsent returns when consent
will be disabled (see https://crbug.com/1041062#c23 for detailed
timeline). By waiting for the device policy file to be present, we skip
over this wait period.

BUG=chromium:1041062
TEST=tast run --extrauseflags chrome_internal $DUT platform.KernelWarning

Change-Id: I17e03915a69607e6e2a1df80c9ee3358f96ba9d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2095752
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>