Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Not sure if it's ready for review given there a many CLs it depends. Please ping me when it's ready.

Thanks for the feedback. It doesn't actually depend on much, I've just been stacking things on top of each other to make uploading and committing all the changes easier. But that probably causes a lot of spam.

I'll make more branches to try to level things a bit.

Patch-set: 7