Introduce ui.EventWatcher and use it in ui.LauncherPageSwitchPerf
There's a TODO for waiting properly on page-switches on launcher.
It turns out that a reliable way of waiting on this context would
be the accessibility event of automation API
https://source.chromium.org/chromium/chromium/src/+/master:ash/app_list/views/page_switcher.cc;l=88

This CL introduces a new entity 'EventWatcher' to chrome/ui package,
which registers itself as a listener of a Node and wait for the
events.

BUG=chromium:1059051
TEST=tast run ui.LauncherPageSwitchPerf.* on krane

Change-Id: I1ad35ebd71b4422af6b8e3563e3351c5826b1fd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2097332
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>