tast/video: Extend shaka.html to automatically change tracks
This CL extends shaka.html to start changing among tracks every
so often if there's more than one present. The interval is set
via parameter but has a default value of 3s (customary DASH).

This CL should not change the behaviour of play.Seekm, since the
dash files used there have only 1 track, but is needed for the
next CL in the chain.

BUG=chromium:1047514
TEST=tast run ${IP} video.Seek.*
  on atlas and grunt.

Change-Id: I3e7f5f17d7ee09d82c577ae850b3d2574e1f5ffe