Switch platform.UserCrash to tast Parameterized test
In https://crbug.com/1060344, we are switching some tast tests to a mock
consent system. We can make mock consent versions of some -- but not all
-- of the tests in platform.UserCrash. The easiest way to track this
will be to parameterize the test suite.

This will also allow us to promote some of these tests to critical even
if not all of them are stable.

BUG=chromium:1060344
TEST=tast on veryon_minnie

Change-Id: I9ce2bbbb390a1a9be58820daa76e12b1104fb7c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2093170
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>