tast-tests: Add ARC input compatibility test
This CL adds arc.InputCompat which tests ARC++ input compatibility
features. In P, this is implemented in the ArcInputCompatHandler class.
There are two such features:

* When the package declares isGame=true, all pointer input will have
  touchscreen input source.
* When the package declares target SDK <= MNC (23), smooth scroll will
  be emulated with two pointers, rather than one pointer.

As the latter requires a virtual trackpad device, this CL implements the
former first. In an upcomping CL, the latter will be implemented as
well.

To test these features, three APKs with identical code except
AndroidManifest.xml are added:
* ArcInputCompatMTest.apk
* ArcInputCompatGameTest.apk
* ArcInputCompatDisabledTest.apk

BUG=b:128546026
TEST=tast run eve-DUT arc.InputCompat (apks should be cros deploy'd
first.)

Change-Id: I7a3bb4b83e3d595bdf532963365f4111d1e287be