Add "gpu_sandboxing" software dep to broker tests
Broker processes are only created if the gpu-process is sandboxed, so
don't look for broker processes on VMs where the gpu-process is not
sandboxed.

Before, this was hidden because metrics_consent and gpu_sandboxing
blocked the same set of boards from the test. But
https://crbug.com/1060344 is removing metrics_consent from many tests,
exposing the need to have an explicit dependency on gpu_sandboxing.

BUG=chromium:1060344
TEST=tast -verbose run --extrauseflags chrome_internal localhost:2222 ui.ChromeCrash*LoggedIn.*; ensured broker tests were not skipped on a veyron_minnie.

Change-Id: I041a421ade9475927e4574b5e3229862d52fcf70