tast/video: speculative Seek.switch_* test fix
crrev.com/c/2076428 simplified the html file used by said tests;
this seems to have broken the chromeboxes (guado, rikku etc) - I
cannot seem to be able to repro and the test is still working on
other e.g. BDW, HSW devices, so I'm thinking this is due to the
removal of the "loop" attribute in said CL, which would prevent
any resolve() coming out of the randomSeek() function, causing
the said Stainless-collected error. This CL is a speculative fix,
by setting the video.loop attribute for seek test cases.

BUG=b:150870698
TEST=tast run IP video.Seek.switch_* on samus and grunt

Change-Id: Id6474875d3aa4b7f1bfd649ddef30e1439aa3a8c