tast/video: speculative Seek.switch_* test fix
crrev.com/c/2076428 simplified the html file used by said tests;
this seems to have broken the chromeboxes (guado, rikku etc) and
the RockChips (scarlet etc) - I was able to repro on scarlet,
basically tracing down to the .onended event being registered
and fired while on seek cases; this CL fixes that by simply
having two methods: playUntilEnd() and play(); the former only
resolve()s on onended fired, while the latter play()s forever.

Note that scarlet for example still fails with only 18/25 seeks
executed.

"Speculative" in the title refers to the fact that I haven't
been able to get my ssh on a rikku, so the fix is developed on
scarlet.

BUG=b:150870698,b:151238010
TEST=tast run IP video.Seek.switch_* video.Play.vp8_hw on samus,
  scarlet

Change-Id: Id6474875d3aa4b7f1bfd649ddef30e1439aa3a8c