tast-test: arc.KillProcess work on VM
arc.KillProcess was failing on ARCVM because the test was querying the
host processes. When ARC runs inside a container, the Android processes
are just "host" processes. But when ARCVM is running, it is not possible
to see the VM processes from the host.

This CL fixes that by parsing the Android processes by using "adb
shell".

BUG=b:150719902
TEST=tast run arc.KillProcess arc.KillProcess.vm

Change-Id: I75efcbacd105a79c86ee4155b7d98192a7072d08