Update patch set 2
Patch Set 2: Code-Review+2

sorry for the delay!

LGTM... I wonder if this test will need to be updated every time that the OOBE dialog is changed... (in any case I'm not sure how frequent is that). 
I'm thinking of once this test gets promoted to CQ, then if the OOBE dialog is changed, this test shouldn't block CQ... not sure whether there is a fail-safe condition. Like a kind of OOBE version, and check for that.

Patch-set: 2
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2