wificell: test_fixture: Support packet capturing on pcap router.
Setup pcap router in test_fixture and spawn capturer when ConfigureAP.

Doing this in TestFixture makes it convenient for tests that just want
some material for debugging. For advanced usage (e.g. parse pcap to
verify the messages), it is also possible for the test to capture the
packets by itself with Router object returned by Pcap() function.

BUG=chromium:1034875
TEST=`tast run $DUT wifi.SimpleConnect*` and check there are pcap files
collected in test results.

Change-Id: I3ef927f537ee2749c6c8020699c0ba13f7d6da6f